# Conflicts:
#	src/knx/group_object.cpp

+++ src/knx/bau27B0.h
@@ -6,9 +6,9 @@
 #include "bau_systemB_device.h"
 #include "rf_medium_object.h"
 #if defined(DeviceFamily_CC13X0)
-  #include "rf_physical_layer_cc1310.h"
+    #include "rf_physical_layer_cc1310.h"
 #else
-  #include "rf_physical_layer_cc1101.h"
+    #include "rf_physical_layer_cc1101.h"
 #endif
 #include "rf_data_link_layer.h"
 #include "cemi_server.h"
@@ -16,32 +16,32 @@
 
 class Bau27B0 : public BauSystemBDevice
 {
-  public:
-    Bau27B0(Platform& platform);
-    void loop() override;
-    bool enabled() override;
-    void enabled(bool value) override;
+    public:
+        Bau27B0(Platform& platform);
+        void loop() override;
+        bool enabled() override;
+        void enabled(bool value) override;
 
-    RfDataLinkLayer* getDataLinkLayer();
-  protected:
-    InterfaceObject* getInterfaceObject(uint8_t idx);
-    InterfaceObject* getInterfaceObject(ObjectType objectType, uint8_t objectInstance);
+        RfDataLinkLayer* getDataLinkLayer();
+    protected:
+        InterfaceObject* getInterfaceObject(uint8_t idx);
+        InterfaceObject* getInterfaceObject(ObjectType objectType, uint16_t objectInstance);
 
-    void doMasterReset(EraseCode eraseCode, uint8_t channel) override;
-  private:
-    RfDataLinkLayer _dlLayer;
-    RfMediumObject _rfMediumObj;
+        void doMasterReset(EraseCode eraseCode, uint8_t channel) override;
+    private:
+        RfDataLinkLayer _dlLayer;
+        RfMediumObject _rfMediumObj;
 #ifdef USE_CEMI_SERVER
-    CemiServer _cemiServer;
-    CemiServerObject _cemiServerObject;
+        CemiServer _cemiServer;
+        CemiServerObject _cemiServerObject;
 #endif
 
-    void domainAddressSerialNumberWriteIndication(Priority priority, HopCountType hopType, const SecurityControl &secCtrl, const uint8_t* rfDoA,
-                                                  const uint8_t* knxSerialNumber) override;
-    void domainAddressSerialNumberReadIndication(Priority priority, HopCountType hopType, const SecurityControl &secCtrl, const uint8_t* knxSerialNumber) override;
-    void individualAddressSerialNumberReadIndication(Priority priority, HopCountType hopType, const SecurityControl &secCtrl, uint8_t* knxSerialNumber) override;
-    void domainAddressSerialNumberWriteLocalConfirm(Priority priority, HopCountType hopType, const SecurityControl &secCtrl, const uint8_t* rfDoA,
-                                                    const uint8_t* knxSerialNumber, bool status) override;
-    void domainAddressSerialNumberReadLocalConfirm(Priority priority, HopCountType hopType, const SecurityControl &secCtrl, const uint8_t* knxSerialNumber, bool status) override;
+        void domainAddressSerialNumberWriteIndication(Priority priority, HopCountType hopType, const SecurityControl& secCtrl, const uint8_t* rfDoA,
+                const uint8_t* knxSerialNumber) override;
+        void domainAddressSerialNumberReadIndication(Priority priority, HopCountType hopType, const SecurityControl& secCtrl, const uint8_t* knxSerialNumber) override;
+        void individualAddressSerialNumberReadIndication(Priority priority, HopCountType hopType, const SecurityControl& secCtrl, uint8_t* knxSerialNumber) override;
+        void domainAddressSerialNumberWriteLocalConfirm(Priority priority, HopCountType hopType, const SecurityControl& secCtrl, const uint8_t* rfDoA,
+                const uint8_t* knxSerialNumber, bool status) override;
+        void domainAddressSerialNumberReadLocalConfirm(Priority priority, HopCountType hopType, const SecurityControl& secCtrl, const uint8_t* knxSerialNumber, bool status) override;
 };
 #endif
